Added `recovered_agents` in `GetAgents` response.
This patch added infos of recovered but yet-to-reregister agents in the
\`GetAgents\` response. This can be used by an operator or subscriber
to know these agents are in this transitional state.
One use case is to perform special handling on previously known
tasks from these agents because these tasks are unknown to master yet.

Bugs closed: MESOS-6177

Reviewed at https://reviews.apache.org/r/52637/