Make `slaves.recovered` in master a hashmap.
This information will be used to in both
\`GetAgents\` API call and \`/state\` endpoint
to return \`AgentInfo\` of recovered but unregistered
slave, to facilitate operator to better handle
such case.

The reason I want to return \`AgentInfo\` instead
of \`AgentID\` is that hostname and ip address
can be find in former, but not latter.

Reviewed at https://reviews.apache.org/r/53095/