Fixed an issue in the gzip error handling.
It turns out that zlib does not always set the `msg` field, the
calling code is expected to handle the case where `msg` is NULL.
I discovered this while I was playing with the library during
the implementation of a streaming decompressor.

To resolve this issue, this patch introduces an GzipError,
currently hidden to callers within an internal namespace,
in order to produce error messages in a consistent (safe)
manner.

In addition, this patch moves the (in|de)flateInit2 and
(in|de)flateEnd failures to ABORT the program, since these
correspond to programming errors and we will need to invoke
these constructors and destructors once we have streaming
(de)compression support.

Review: https://reviews.apache.org/r/53365