Added GPU allocation in the docker containerizer.
This patch allocates and de-allocates GPU devices during
container launch in the docker containerizer. It does not
yet support updating resources, recovering these resources
after agent restart, or passing them to 'docker' itself.
In other words, this only adds some (but not all!) of the
steady-state accounting of the GPU resources; the GPUs are
not yet available to the containers.

Note that the behavior of the docker containerizer before
this patch is that GPU resources are ignored entirely.

Review: https://reviews.apache.org/r/50841/