Used 0755 instead of 0750 for sandbox permission.
Summary:
Due to issues of "de-escalate" from root to udocker in our entrypoint
script, we want to revert this to previous behavior which uses 0755 for
sandbox permission.

This is a uber specific patch for now since upstream has not signed off
a full revert yet.

Test Plan: make

Reviewers: jlai,echung

Subscribers:

Revert Plan: